# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations/{organizationId}/feature-flags": {
      "get": {
        "operationId": "OperationsController_listFlags",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List feature flags resolved for the authenticated user",
        "tags": ["Operations"]
      }
    },
    "/api/v1/organizations/{organizationId}/feature-flags/{key}": {
      "put": {
        "operationId": "OperationsController_setFlag",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set an organization or user feature flag",
        "tags": ["Operations"]
      }
    },
    "/api/v1/organizations/{organizationId}/audit": {
      "get": {
        "operationId": "OperationsController_audit",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Query append-only audit events for the organization",
        "tags": ["Operations"]
      }
    },
    "/api/v1/organizations/{organizationId}/jobs/{queue}": {
      "post": {
        "operationId": "OperationsController_enqueue",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "queue",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enqueue a bounded background job",
        "tags": ["Operations"]
      }
    },
    "/api/v1/organizations/{organizationId}/jobs/{queue}/dead-letters": {
      "get": {
        "operationId": "OperationsController_deadLetters",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "queue",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Inspect dead-lettered jobs for the organization",
        "tags": ["Operations"]
      }
    },
    "/api/v1/organizations/{organizationId}/jobs/{queue}/dead-letters/{idempotencyKey}/retry": {
      "post": {
        "operationId": "OperationsController_retryDeadLetter",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "queue",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotencyKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retry one dead-lettered job",
        "tags": ["Operations"]
      }
    },
    "/api/v1/organizations/{organizationId}/jobs/{queue}/dead-letters/{idempotencyKey}": {
      "delete": {
        "operationId": "OperationsController_discardDeadLetter",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "queue",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotencyKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Discard one dead-lettered job with audit",
        "tags": ["Operations"]
      }
    },
    "/api/v1/operations": {
      "post": {
        "operationId": "AsyncOperationsController_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an organization-scoped asynchronous operation",
        "tags": ["Operations"]
      }
    },
    "/api/v1/operations/{id}": {
      "get": {
        "operationId": "AsyncOperationsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read an organization-scoped asynchronous operation",
        "tags": ["Operations"]
      }
    },
    "/api/v1/operations/{id}/cancel": {
      "post": {
        "operationId": "AsyncOperationsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel an organization-scoped asynchronous operation",
        "tags": ["Operations"]
      }
    },
...
  },
  "info": {
    "title": "HandStack API",
    "description": "Versioned public and administrative HandStack API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}